+++ apps/trading/components/accounts-container/accounts-container.tsx
@@ -2,19 +2,9 @@ import { useNavigate, useSearchParams } from 'react-router-dom';
 import { useAssetDetailsDialogStore } from '@vegaprotocol/assets';
 import { useVegaWallet } from '@vegaprotocol/wallet-react';
 import { Links } from '../../lib/links';
-import {
-  AccountManager,
-  type AssetActions,
-  TransferContainer,
-} from '@vegaprotocol/accounts';
-import { useState } from 'react';
-import { VegaIcon, VegaIconNames } from '@vegaprotocol/ui-toolkit';
-import { useT } from '../../lib/use-t';
-import { DepositContainer } from '@vegaprotocol/deposits';
-import { SwapContainer } from '../swap';
-import { WithdrawContainer } from '../withdraw-container';
+import { AccountManager, type AssetActions } from '@vegaprotocol/accounts';
 
-type AccountsContainerProps = Partial<AssetActions> & {
+export type AccountsContainerProps = Partial<AssetActions> & {
   pinnedAssets?: string[];
   orderByBalance?: boolean;
   hideZeroBalance?: boolean;
@@ -74,88 +64,6 @@ export const AccountsContainer = ({
       pinnedAssets={pinnedAssets}
       orderByBalance={orderByBalance}
       hideZeroBalance={hideZeroBalance}
-    />
-  );
-};
-
-enum View {
-  Deposit = 'Deposit',
-  Swap = 'Swap',
-  Transfer = 'Transfer',
-  Withdraw = 'Withdraw',
-}
-
-type InnerView = [view: View, assetId: string];
-
-export const SidebarAccountsContainer = ({
-  pinnedAssets,
-}: Pick<AccountsContainerProps, 'pinnedAssets'>) => {
-  const t = useT();
-  const [innerView, setInnerView] = useState<InnerView | undefined>(undefined);
-
-  if (innerView) {
-    const [view, assetId] = innerView;
-
-    let innerContainer = undefined;
-    switch (view) {
-      case View.Deposit:
-        innerContainer = <DepositContainer assetId={assetId} />;
-        break;
-      case View.Swap:
-        innerContainer = (
-          <SwapContainer
-            assetId={assetId}
-            onDeposit={(assetId) => {
-              if (!assetId) return;
-              setInnerView([View.Deposit, assetId]);
-            }}
-          />
-        );
-        break;
-      case View.Transfer:
-        innerContainer = <TransferContainer assetId={assetId} />;
-        break;
-      case View.Withdraw:
-        innerContainer = <WithdrawContainer assetId={assetId} />;
-    }
-
-    return (
-      <div className="p-2">
-        <header className="text-lg">
-          <button
-            className="flex gap-1 items-center"
-            onClick={() => {
-              setInnerView(undefined);
-            }}
-          >
-            <span className="text-muted inline-flex items-center">
-              <VegaIcon name={VegaIconNames.CHEVRON_LEFT} />
-            </span>
-            <span>{t(view)}</span>
-          </button>
-        </header>
-        <div className="mt-2">{innerContainer}</div>
-      </div>
-    );
-  }
-
-  return (
-    <AccountsContainer
-      pinnedAssets={pinnedAssets}
-      orderByBalance
-      hideZeroBalance
-      onClickDeposit={(assetId) => {
-        setInnerView([View.Deposit, assetId]);
-      }}
-      onClickSwap={(assetId) => {
-        setInnerView([View.Swap, assetId]);
-      }}
-      onClickTransfer={(assetId) => {
-        setInnerView([View.Transfer, assetId]);
-      }}
-      onClickWithdraw={(assetId) => {
-        setInnerView([View.Withdraw, assetId]);
-      }}
     />
   );
 };
