+++ apps/trading/components/accounts-container/index.ts
@@ -1,1 +1,2 @@
 export * from './accounts-container';
+export * from './sidebar-accounts-container';
