+++ apps/trading/components/sidebar/sidebar.tsx
@@ -14,7 +14,7 @@ import { NodeHealthContainer } from '../node-health';
 import { AssetCard } from '../asset-card';
 import { Links } from '../../lib/links';
 import { useT } from '../../lib/use-t';
-import { SidebarAccountsContainer } from '../accounts-container';
+import { SidebarAccountsContainer } from '../accounts-container/sidebar-accounts-container';
 import classNames from 'classnames';
 
 export enum ViewType {
@@ -80,6 +80,7 @@ export const Sidebar = ({ pinnedAssets }: { pinnedAssets?: string[] }) => {
           ) : (
             <AccordionPrimitive.Header>
               <AccordionPrimitive.Trigger
+                data-testid="Assets"
                 className={classNames('grid w-full', {
                   'grid-cols-2': pinnedAssets.length === 2,
                   'grid-cols-1': pinnedAssets.length === 1,
